Changed the data type of tranAmt from Integer to BigDecimal. And added transactions list in final response.

+++ src/main/java/com/maymar/rewards/program/entity/CustomerTransactionsEntity.java
@@ -23,7 +23,7 @@ public class CustomerTransactionsEntity {
 
     private String userId;
 
-    private Integer tranAmt;
+    private BigDecimal tranAmt;
 
     private LocalDate tranDate;
 }
